Fix item damage show in client
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@1356 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/ItemPrototype.h
@@ -162,8 +162,8 @@ enum ITEM_SUBCLASS_PROJECTILE
 
 struct _Damage
 {
-    uint32 DamageMin;
-    uint32 DamageMax;
+    float DamageMin;
+    float DamageMax;
     uint32 DamageType;
 
 };
